requirements.txt
# Core NLP libraries
spacy>=3.4.0
pandas>=1.5.0

# Document processing
python-docx>=0.8.11
PyPDF2>=3.0.0

# OCR support (optional)
pytesseract>=0.3.10
Pillow>=9.0.0
pdf2image>=1.16.0

# Data processing
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0

# For downloading the dataset
kagglehub>=0.2.0
